+++ src/app/pages/home-admin/home-admin.component.spec.ts
@@ -1,21 +1,16 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { Component, OnInit } from '@angular/core';
+import { AuthService } from 'src/app/services/auth.service';
+import { ActivatedRoute, Router } from '@angular/router';
+import { AlertService } from 'src/app/services/alert.service';
+import { IError } from 'src/shared/model/http-error.model';
+import { HttpErrorResponse } from '@angular/common/http';
 
-import { HomeAdminComponent } from './home-admin.component';
+type ErrorResponseType = HttpErrorResponse;
 
-describe('HomeAdminComponent', () => {
-  let component: HomeAdminComponent;
-  let fixture: ComponentFixture<HomeAdminComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [HomeAdminComponent]
-    });
-    fixture = TestBed.createComponent(HomeAdminComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+@Component({
+  selector: 'app-home-admin',
+  templateUrl: './home-admin.component.html',
+  styleUrls: ['./home-admin.component.css']
+})
+export class HomeAdminComponent {
+}
